Fixing some bugs:
In find_seq_head now I correctly identify the root of the sentence as
the head of the subsequence only if it is inside the subsequence instead

In is_isolated I now ignore outward arcs towards punctuation in the
count. This was a mistake, as I am only interested in actual words.

Changes to be committed:
modified:   computation/app/src/extract.py